# pkg/server/route/docs/swagger.yaml
definitions:
...
  cluster_token.HttpReq_ClusterToken_create:
    properties:
      cluster_uuid:
        type: string
      name:
        type: string
      summary:
        type: string
      uuid:
        description: optional
        type: string
    type: object
  cluster_token.HttpReq_ClusterToken_update:
    properties:
      name:
        type: string
      summary:
        type: string
    type: object
  cluster_token.HttpRsp_ClusterToken:
    properties:
      cluster_uuid:
        type: string
      created:
        type: string
      deleted:
        type: string
      expiration_time:
        type: string
      id:
        description: pk
        type: integer
      issued_at_time:
        type: string
      name:
        type: string
      summary:
        type: string
      token:
        type: string
      updated:
        type: string
      uuid:
        description: uuid
        type: string
    type: object
...
info:
  contact:
    email: jaehoon@nexclipper.io
    url: https://nexclipper.io
  description: this is a sudory server.
  title: SUDORY
  version: 0.0.1
paths:
...
  /server/cluster_token:
    get:
      consumes:
      - application/json
      description: Find Cluster Token
      parameters:
      - description: query  github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: q
        type: string
      - description: order  github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: o
        type: string
      - description: paging github.com/NexClipper/sudory/pkg/server/database/vanilla/stmt/README.md
        in: query
        name: p
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cluster_token.HttpRsp_ClusterToken'
            type: array
      security:
      - ServiceAuth: []
      tags:
      - server/cluster_token
    post:
      consumes:
      - application/json
      description: Create a Cluster Token
      parameters:
      - description: ClusterToken HttpReq_ClusterToken_create
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/cluster_token.HttpReq_ClusterToken_create'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cluster_token.HttpRsp_ClusterToken'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster_token
  /server/cluster_token/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete a Cluster Token
      parameters:
      - description: ClusterToken Uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ServiceAuth: []
      tags:
      - server/cluster_token
    get:
      consumes:
      - application/json
      description: Get a Cluster Token
      parameters:
      - description: ClusterToken Uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cluster_token.HttpRsp_ClusterToken'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster_token
  /server/cluster_token/{uuid}/expire:
    put:
      consumes:
      - application/json
      description: Expire Cluster Token
      parameters:
      - description: ClusterToken Uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cluster_token.HttpRsp_ClusterToken'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster_token
  /server/cluster_token/{uuid}/label:
    put:
      consumes:
      - application/json
      description: Update Label of Cluster Token
      parameters:
      - description: ClusterToken Uuid
        in: path
        name: uuid
        required: true
        type: string
      - description: ClusterToken HttpReq_ClusterToken_update
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/cluster_token.HttpReq_ClusterToken_update'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cluster_token.HttpRsp_ClusterToken'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster_token
  /server/cluster_token/{uuid}/refresh:
    put:
      consumes:
      - application/json
      description: Refresh Time of Cluster Token
      parameters:
      - description: ClusterToken Uuid
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cluster_token.HttpRsp_ClusterToken'
      security:
      - ServiceAuth: []
      tags:
      - server/cluster_token
...
securityDefinitions:
  ClientAuth:
    description: token for client api
    in: header
    name: x-sudory-client-token
    type: apiKey
  ServiceAuth:
    description: Bearer token for service api
    in: header
    name: Authorization
    type: apiKey
  XAuthToken:
    description: limit for access sudory api
    in: header
    name: x_auth_token
    type: apiKey
swagger: "2.0"